小数点以下を切り捨てない

+++ src/main/kotlin/Calculator.kt
@@ -3,5 +3,5 @@ package org.example
 class Calculator {
     fun multiply(x: Int, y: Int): Int = x * y
 
-    fun divide(x: Int, y: Int): Int = x / y
+    fun divide(x: Int, y: Int): Float = x.toFloat() / y.toFloat()
 }
